requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
Flask-Limiter==2.10.4
bcrypt==4.0.1
email-validator==1.3.1
python-dotenv==1.0.0
reportlab==4.0.0
gunicorn==21.0.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
